requirements.txt
openai
python-dotenv
pillow
